requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2022.1
Django==4.2.5
django-crispy-forms==2.0
dominate==2.8.0
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
pdf2image==1.16.3
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
visitor==0.1.3
Werkzeug==2.3.7